Extract TimeframeDropdown into a dedicated component
