+++ components/index.ts
@@ -12,5 +12,6 @@ export { default as SalesTable } from './SalesTable'
 export { default as StockTable } from './StockTable'
 export { default as Table } from './Table'
 export { default as TextInput } from './TextInput'
+export { default as TimeframeDropdown } from './TimeframeDropdown'
 export { default as ValidatedDropdown } from './ValidatedDropdown'
 export * from './employees'
